update repository
Forgot I had this on git, way too much has changed to fit in a commit description

+++ src/firemerald/mcms/api/util/RaytraceResult.java
@@ -5,9 +5,9 @@ import firemerald.mcms.api.model.IRaytraceTarget;
 public class RaytraceResult
 {
 	public final IRaytraceTarget hit;
-	public final float m;
+	public final double m;
 	
-	public RaytraceResult(IRaytraceTarget hit, float m)
+	public RaytraceResult(IRaytraceTarget hit, double m)
 	{
 		this.hit = hit;
 		this.m = m;
